update log impl
