+++ compiler/codegen/codegen.py
@@ -1,14 +1,5 @@
-def begin_sep(sec):
-    return f"\n///@@ BEG_OF {sec} @@\n"
-
-def end_sep(sec):
-    return f"\n///@@ END_OF {sec} @@\n"
-
-def to_str_debug(x):
-    return f'format!("{{:?}}", {x})'
-
-def to_str(x):
-    return f'format!("{{}}", {x})'
+from codegen.helper import *
+from codegen.snippet import *
 
 def compile_sql_to_rust(ast, ctx):
     if ast["type"] == "CreateTableAsStatement":
@@ -30,81 +21,16 @@ def compile_sql_to_rust(ast, ctx):
         print(ast)
         raise ValueError("Unsupported SQL statement")
 
-def type_mapping(sql_type):
-    if sql_type == "TIMESTAMP":
-        return "DateTime<Utc>"
-    elif sql_type == "VARCHAR":
-        return "String"
-    else:
-        raise ValueError("Unknown type")
-
-def input_mapping(fields):
-    if fields == "CURRENT_TIMESTAMP":
-        return "Utc::now()"
-    elif fields == "event_type":
-        return to_str_debug("meta_ref.msg_type")
-    elif fields == "src":
-        return to_str_debug("meta_ref.conn_id")
-    elif fields == "dst":
-        return to_str_debug("meta_ref.conn_id")
-    elif fields == "rpc":
-        return to_str("req.addr_backend.clone()")
-    elif fields == "meta_buf_ptr":
-        return "req.meta_buf_ptr.clone()"
-    elif fields == "addr_backend":
-        return "req.addr_backend.clone()"
-    else:
-        print(fields)
-        raise ValueError("Unknown field")
-
 def handle_create_table_statement(ast, ctx):
     table_name = ast["table_name"]
     if table_name == "output":
         raise ValueError("Table name 'output' is reserved")
-    vec_name = "table_" + ast["table_name"]
-    struct_name = "struct_" + ast["table_name"]
-    table = {
-        "name": vec_name,
-        "type": "Vec",
-        "struct": {
-            "name": struct_name,
-            "fields": []
-        },
-    }
-    if ctx["tables"].get(table_name) is None:
-        ctx["tables"][table_name] = table
+    if table_name.endswith("_file"):
+        ret = generate_create_for_file(ast, ctx, table_name)
     else:
-        raise ValueError("Table already exists")
-    
-    rust_struct = f"pub struct {struct_name} {{\n"
-    for column in ast["columns"]:
-        rust_type = type_mapping(column["data_type"])
-        table["struct"]["fields"].append({"name": column["column_name"], "type": rust_type})
-        rust_struct += f"    pub {column['column_name']}: {rust_type},\n"
-    rust_struct += "}\n"
-    
-    rust_impl = f"impl {struct_name} {{\n"
-    rust_impl += f"     pub fn new("
-    for column, idx in zip(ast["columns"], range(len(ast["columns"]))):
-        rust_impl += f"{column['column_name']}: {type_mapping(column['data_type'])}"
-        if idx != len(ast["columns"]) - 1:
-            rust_impl += ", "
-    rust_impl += f") -> {struct_name} {{\n"
-    
-    rust_impl += f"         {struct_name} {{\n"
-    for column in ast["columns"]:
-        rust_impl += f"             {column['column_name']}: {column['column_name']},\n"
-    rust_impl += f"         }}\n"
-    rust_impl += f"     }}\n"
-    rust_impl += f"}}\n"
-    
-    rust_vec = begin_sep("name") + f"{vec_name}" + end_sep("name") 
-    rust_vec += begin_sep("type") + f"{table['type']}<{struct_name}>" + end_sep("type")
-    rust_vec += begin_sep("init") + f"{vec_name} = Vec::new()" + end_sep("init")
-
-    rust_internal = begin_sep("internal") + f"pub {vec_name}: Vec<{struct_name}>," + end_sep("internal")
-    return begin_sep("declaration") + rust_struct + "\n" + rust_impl + "\n" + end_sep("declaration") + "\n" + rust_vec + "\n" + rust_internal + "\n" + begin_sep("process")
-
+        ret = generate_create_for_vec(ast, ctx, table_name)
+    return ret
+ 
 def handle_insert_statement(node, ctx):
     table_name = node["table_name"]
     table = ctx["tables"].get(table_name)
@@ -117,7 +43,14 @@ def handle_insert_statement(node, ctx):
         select = value
         select["to"] = table_name
         select_statement = handle_select_statement(select, ctx)
-        return f"for event in {select_statement} {{ {vec_name}.push(event); }}"
+        rust_code = f"for event in {select_statement} {{"
+        if table_name.endswith("file"):
+            file_name = table["file_field"]
+            rust_code += f"write!(self.{file_name}, \"{{}}\", {vec_name});"
+        else:
+            rust_code += f"{vec_name}.push(event);"
+        rust_code += f"}}"
+        return rust_code
     else:
         codes = ""
         struct_name = table["struct"]["name"]
